Restructured fan performance data
For #209. This restructuring is required to avoid that mover data require the user to enter data that are not used by the model. Using default values such as V_flow={0} does not work as this lead to a division by zero in some model. Also, not declaring a default value also does not work as this led to error such as 'Failed to expand the variable floMacSta.per.pressure.V_flow'. Therefore, the package has been redesigned so that only the minimum set of required data need to be entered. See also the updated user's guide.