Set tau=0 for enthalpy sensors, with final attribute.
Added <code>final</code> to <code>tau = 0</code> in
<code>EnthalpyFlowRate</code>.
These sensors do not need dynamics as the enthalpy flow rate
is used to compute a heat flow which is then added to the volume of the
tank.
Thus, if there were high frequency oscillations of small mass flow
rates,
then they have a small effect on <code>H_flow</code>, and they are
not used in any control loop. Rather, the oscillations are further
damped
by the differential equation of the fluid volume.