Removed Manifold example
These files were removed for #426.
Dymola 2016 seems to have issues when checking whether start values of
alias variables have the same value if they are the result of a
computation. The other heat exchangers that use the manifold check and
translate correctly, but this example fails, maybe because there are no
volumes between the manifolds.