Update log functions to avoid using ModelicaUtilities
The Modelica log utility functions use global state that is not in scope
or otherwise not initialized. Avoid this by using "EnergyPlus" equivalents,
thereby saving the task of feeding energyplus log statements through the
Modelica logging system, for some point in the future.

The new file EnergyPlusModelicaUtilities.c has been added to the CMake build rules