Refactored time tables to avoid events
This removes any zero crossing functions. Also the offset has been removed for the integer and the boolean time table, and the time table has been redesigned. Previously, it required a time stamp for the last instant in the period, but its values are never used, hence this was refactored