Made BaseProperties replaceable in Media.Air (#2589)
* Made BaseProperties replaceable in Media.Air

For IBPSA, PR 1516

* Merged IBPSA, patch_air_replaceable_baseproperties

* Refactored energy output calculation and air infiltration

This commit replaces the heating or cooling power output of the room with outputs for heating or cooling energy. This is done because the values quickly change, and in the CI testing, this output variable caused often a failure at the very last time stamp of the simulation.

This version also improves the infiltration. Now, exactly the same amount of air in infiltrated and
exfiltrated. This was not the case previously because the infiltration was a prescribed air flow rate,
and the exfiltration was based on pressure difference. This caused an inbalance in the HVAC supply and
return air flow rate.

* Excluded simulation with JModelica, see https://app.travis-ci.com/github/lbl-srg/modelica-buildings/jobs/531423041