Integrated IBPSA, issue1477_diffuseRadiation and issue1478_BESTEST_weatherfiles
BESTEST and other results only changed slightly. Some examples such as Buildings.HeatTransfer.Windows.BaseClasses.Examples.WindowRadiation changed significantly because the model used the latitude as a parameter which was set to a different value than the location of the weather data file has. With this code change, the latitude from the weather data file is used