Merge new weather data implementation from IBPSA (#2616)
* Updated release notes and conversion script

* Merged IBPSA, issue1477_diffuseRadiation_AND_issue1478_BESTEST_weatherfiles and issue1494_plugflowpipe_singleoutlet

BESTEST and other results only changed slightly. Some examples such as Buildings.HeatTransfer.Windows.BaseClasses.Examples.WindowRadiation changed significantly because the model used the latitude as a parameter which was set to a different value than the location of the weather data file has. With this code change, the latitude from the weather data file is used

* Updated BESTEST results

* Added input signal for altitude

Dymola assumes a signal of 0 if it is not set on the weather data bus, whereas OCT issues a translation error