Issue2725 boiler heat correction (#2748)
* improved the usersguide

In preparation, the usersguide is partially rewrote so that the description of equations correspond to the model text equation by equation. Some explanatory texts are also rearranged to improve clarity. Some redundant or duplicate documentation left behind in #2695 is cleaned out.

* implemented boiler correction

The changes are very small and I suspect that it was caused by my using Ubuntu 20 instead of 18. But I am pushing it anyway so that it could be tested by someone else.

* Revised documentation, removed trailing white spaces

* Removed trailing white spaces

Done by running for ff in Fluid/Boilers/*; do sed -e 's/ *$//g' -i Fluid/Boilers/BoilerPolynomial.mo; done;

Co-authored-by: Michael Wetter <mwetter@lbl.gov>