reinstated check valve in NetworkConnection
The check valves were removed because of numerical issues and control logics were designed to prevent substantial back flow at the supply pump. However, these logics have been found to be incompatible with the main control objectives. In the meantime, because the numerics of the models have been substantially improved with extensive use of junction models, the check valve no longer causes OPTIMICA to fail simulation. This is then thought to be a simpler solution than convoluted control logic.