Issue3479 misc fix (#3480)
* fixed typo in Buildings.Fluid.FMI.UsersGuide

* improved mover model assert messages

* moved connection line to void overlapping

The connection line out of returnAir is moved because it overlapped with many other lines. Also slightly moved the fan to straighten connection lines. Slightly moved another connection line out of the chiller so that it is not so close to another line.

* fixed typo in equation

Corrected an equation in documentaiton where dT should be T. Also rewrote a sentence for clarity. And fixed another minor documentation typo.

* fixed typo

* fixed broken mos that caused missing figure

* corrected mos to remove empty subplot